[network-tracer] Fix C.uint -> uint type casting for go version prior to 1.11 and update Dockerfile-nettop go version to 1.11.2

+++ ebpf/tracer.go
@@ -204,7 +204,7 @@ func loadBPFModule() (*bpflib.Module, error) {
 }
 
 func (t *Tracer) timeoutForConn(c *ConnTuple) int64 {
-	if connType(c.metadata) == TCP {
+	if connType(uint(c.metadata)) == TCP {
 		return t.config.TCPConnTimeout.Nanoseconds()
 	}
 	return t.config.UDPConnTimeout.Nanoseconds()
